consolidate token bottom sheets
