+++ src/components/TokenBottomSheet.test.tsx
@@ -5,7 +5,7 @@ import { Provider } from 'react-redux'
 import { TokenBottomSheetEvents } from 'src/analytics/Events'
 import ValoraAnalytics from 'src/analytics/ValoraAnalytics'
 import TokenBottomSheet from 'src/components/TokenBottomSheet'
-import { DEBOUCE_WAIT_TIME, TokenPickerOrigin } from 'src/components/TokenBottomSheetByAddress'
+import { DEBOUCE_WAIT_TIME, TokenPickerOrigin } from 'src/components/TokenBottomSheet'
 import { TokenBalance } from 'src/tokens/slice'
 import { NetworkId } from 'src/transactions/types'
 import { createMockStore, getElementText } from 'test/utils'
